server_py/pyproject.toml
[project]
name = "edtech-lms-api"
version = "1.0.0"
description = "EdTech LMS FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "itsdangerous>=2.2.0",
]
